requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pymongo==4.6.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.3
PyPDF2==3.0.1
python-multipart==0.0.6
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0
